requirements.txt
APScheduler==3.10.4
argcomplete==3.2.2
beautifulsoup4==4.9.3
blinker==1.7.0
boto3==1.34.35
botocore==1.34.35
bs4==0.0.2
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.7
config==0.5.1
cssselect==1.2.0
dnspython==1.16.0
durationpy==0.6
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.2
Flask-APScheduler==1.13.1
gnews==0.3.6
hjson==3.1.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kappa==0.6.0
lxml==5.1.0
MarkupSafe==2.1.5
newspaper3k==0.2.8
nltk==3.8.1
pillow==10.2.0
placebo==0.9.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.26.0
requests-file==2.0.0
s3transfer==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
text-unidecode==1.3
tinysegmenter==0.3
tldextract==5.1.1
toml==0.10.2
tqdm==4.66.1
troposphere==4.6.0
tzlocal==5.2
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
